Support for RO replicas in replica-install
ipa-replica-install now supports read-only replicas. New options were
added: '--consumer' - to specify Consumer type of replica, '--hub' - to
specify Hub type of replica and '--shadow' to set the Hub replica to be
hidden in the network topology. For shadow replica no Kerberos and WebUI
is provided. Read-only replicas also do not support CA to be installed.